Closes #75 #110
