+++ common/src/main/java/com/microsoft/identity/common/internal/dto/RefreshToken.java
@@ -27,7 +27,6 @@ import com.google.gson.annotations.SerializedName;
 import static com.microsoft.identity.common.internal.dto.RefreshToken.SerializedNames.CLIENT_INFO;
 import static com.microsoft.identity.common.internal.dto.RefreshToken.SerializedNames.FAMILY_ID;
 import static com.microsoft.identity.common.internal.dto.RefreshToken.SerializedNames.TARGET;
-import static com.microsoft.identity.common.internal.dto.RefreshToken.SerializedNames.USERNAME;
 
 
 public class RefreshToken extends Credential {
@@ -45,10 +44,6 @@ public class RefreshToken extends Credential {
          * String of target.
          */
         public static final String TARGET = "target";
-        /**
-         * String of username.
-         */
-        public static final String USERNAME = "username";
     }
 
     /**
@@ -72,14 +67,6 @@ public class RefreshToken extends Credential {
      */
     @SerializedName(TARGET)
     private String mTarget;
-
-    /**
-     * The primary username that represents the user (corresponds to the preferred_username claim
-     * in the v2.0 endpoint). It could be an email address, phone number, or a generic username
-     * without a specified format. Its value is mutable and might change over time.
-     */
-    @SerializedName(USERNAME)
-    private String mUsername;
 
     /**
      * Gets the target.
@@ -135,62 +122,42 @@ public class RefreshToken extends Credential {
         mFamilyId = familyId;
     }
 
-    /**
-     * Gets the username.
-     *
-     * @return The username to get.
-     */
-    public String getUsername() {
-        return mUsername;
-    }
-
-    /**
-     * Sets the username.
-     *
-     * @param username The username to set.
-     */
-    public void setUsername(final String username) {
-        mUsername = username;
-    }
-
+    //CHECKSTYLE:OFF
     @Override
     public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-
-        if (!super.equals(o)) {
-            return false;
-        }
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        if (!super.equals(o)) return false;
 
         RefreshToken that = (RefreshToken) o;
 
-        if (mClientInfo != null ? !mClientInfo.equals(that.mClientInfo) : that.mClientInfo != null) {
+        if (mClientInfo != null ? !mClientInfo.equals(that.mClientInfo) : that.mClientInfo != null)
             return false;
-        }
-
-        if (mFamilyId != null ? !mFamilyId.equals(that.mFamilyId) : that.mFamilyId != null) {
+        if (mFamilyId != null ? !mFamilyId.equals(that.mFamilyId) : that.mFamilyId != null)
             return false;
-        }
-
-        if (mTarget != null ? !mTarget.equals(that.mTarget) : that.mTarget != null) {
-            return false;
-        }
-
-        return mUsername != null ? mUsername.equals(that.mUsername) : that.mUsername == null;
+        return mTarget != null ? mTarget.equals(that.mTarget) : that.mTarget == null;
     }
+    //CHECKSTYLE:ON
 
+    //CHECKSTYLE:OFF
     @Override
     public int hashCode() {
         int result = super.hashCode();
-        result = UNIQUE_ID_LENGTH * result + (mClientInfo != null ? mClientInfo.hashCode() : 0);
-        result = UNIQUE_ID_LENGTH * result + (mFamilyId != null ? mFamilyId.hashCode() : 0);
-        result = UNIQUE_ID_LENGTH * result + (mTarget != null ? mTarget.hashCode() : 0);
-        result = UNIQUE_ID_LENGTH * result + (mUsername != null ? mUsername.hashCode() : 0);
+        result = 31 * result + (mClientInfo != null ? mClientInfo.hashCode() : 0);
+        result = 31 * result + (mFamilyId != null ? mFamilyId.hashCode() : 0);
+        result = 31 * result + (mTarget != null ? mTarget.hashCode() : 0);
         return result;
     }
+    //CHECKSTYLE:ON
+
+    //CHECKSTYLE:OFF
+    @Override
+    public String toString() {
+        return "RefreshToken{" +
+                "mClientInfo='" + mClientInfo + '\'' +
+                ", mFamilyId='" + mFamilyId + '\'' +
+                ", mTarget='" + mTarget + '\'' +
+                "} " + super.toString();
+    }
+    //CHECKSTYLE:ON
 }
