+++ common/src/main/java/com/microsoft/identity/common/internal/providers/microsoft/azureactivedirectory/AzureActiveDirectoryRefreshToken.java
@@ -24,24 +24,9 @@ package com.microsoft.identity.common.internal.providers.microsoft.azureactivedi
 
 import android.support.annotation.NonNull;
 
-import com.microsoft.identity.common.adal.internal.util.StringExtensions;
-import com.microsoft.identity.common.exception.ServiceException;
-import com.microsoft.identity.common.internal.cache.SchemaUtil;
-import com.microsoft.identity.common.internal.providers.oauth2.IDToken;
-import com.microsoft.identity.common.internal.providers.oauth2.RefreshToken;
+import com.microsoft.identity.common.internal.providers.microsoft.MicrosoftRefreshToken;
 
-import java.util.Date;
-import java.util.concurrent.TimeUnit;
-
-public class AzureActiveDirectoryRefreshToken extends RefreshToken {
-
-    private boolean mIsFamilyRefreshToken;
-    private String mFamilyId;
-    private Date mExpiresOn;
-    private ClientInfo mClientInfo;
-    private IDToken mIdToken;
-    private String mScope;
-    private String mClientId;
+public class AzureActiveDirectoryRefreshToken extends MicrosoftRefreshToken {
 
     /**
      * Constructor of AzureActiveDirectoryRefreshToken.
@@ -51,68 +36,6 @@ public class AzureActiveDirectoryRefreshToken extends RefreshToken {
     public AzureActiveDirectoryRefreshToken(
             @NonNull final AzureActiveDirectoryTokenResponse response) {
         super(response);
-        try {
-            mFamilyId = response.getFamilyId();
-            mIsFamilyRefreshToken = !StringExtensions.isNullOrBlank(mFamilyId);
-            mExpiresOn = response.getExpiresOn();
-            mClientInfo = new ClientInfo(response.getClientInfo());
-            mIdToken = new IDToken(response.getIdToken());
-            mScope = response.getScope();
-            mClientId = response.getClientId();
-        } catch (ServiceException e) {
-            // TODO handle this properly
-            throw new RuntimeException(e);
-        }
-    }
-
-    /**
-     * @return true if the token is family refresh token, false otherwise.
-     */
-    public boolean getIsFamilyRefreshToken() {
-        return mIsFamilyRefreshToken;
-    }
-
-    @Override
-    public String getFamilyId() {
-        return mFamilyId;
-    }
-
-    @Override
-    public String getHomeAccountId() {
-        return SchemaUtil.getHomeAccountId(mClientInfo);
-    }
-
-    @Override
-    public String getEnvironment() {
-        return SchemaUtil.getEnvironment(mIdToken);
-    }
-
-    @Override
-    public String getClientId() {
-        return mClientId;
-    }
-
-    @Override
-    public String getSecret() {
-        return getRefreshToken();
-    }
-
-    @Override
-    public String getTarget() {
-        return mScope;
-    }
-
-    @Override
-    public String getExpiresOn() {
-        String expiresOn = null;
-
-        if (null != mExpiresOn) {
-            final long millis = mExpiresOn.getTime();
-            final long seconds = TimeUnit.MILLISECONDS.toSeconds(millis);
-            expiresOn = String.valueOf(seconds);
-        }
-
-        return expiresOn;
     }
 
 }
